+++ biss.py
@@ -34,7 +34,7 @@ Usage:
     biss <command> --help
 
 Author: Dean Thomas (@Deenyoro)
-Version: 2.2.2
+Version: 2.2.3
 """
 
 import sys